+++ src/interfaces/PenalCode.ts
@@ -3,4 +3,5 @@ export interface PenalCode {
   title: string;
   des: string;
   jail_time: string;
+  fine_amount: string;
 }